Add confim modal for deleting todo items in the BlazorDual template #5166

+++ src/Templates/BlazorDual/Bit.BlazorDual/src/BlazorDual.Web/Pages/Todo/TodoPage.razor.cs
@@ -11,6 +11,7 @@ public partial class TodoPage
     private string? _selectedSort;
     private string? _selectedFilter;
     private string? _underEditTodoItemTitle;
+    ConfirmMessageBox _confirmMessageBox = default!;
     private string _newTodoTitle = string.Empty;
     private IList<TodoItemDto> _allTodoItems = default!;
     private IEnumerable<TodoItemDto> _viewTodoItems = default!;
@@ -133,15 +134,21 @@ public partial class TodoPage
     {
         if (_isLoading) return;
 
-        _isLoading = true;
-
         try
         {
-            await HttpClient.DeleteAsync($"TodoItem/Delete/{todoItem.Id}");
+            var confirmed = await _confirmMessageBox.Show(Localizer.GetString(nameof(AppStrings.AreYouSureWannaDelete), todoItem.Title ?? string.Empty),
+                                                     Localizer[nameof(AppStrings.DeleteTodoItem)]);
 
-            _allTodoItems.Remove(todoItem);
+            if (confirmed)
+            {
+                _isLoading = true;
 
-            FilterViewTodoItems();
+                await HttpClient.DeleteAsync($"TodoItem/Delete/{todoItem.Id}");
+
+                _allTodoItems.Remove(todoItem);
+
+                FilterViewTodoItems();
+            }
         }
         finally
         {
